fix(lint): silence all clang-tidy warnings
.clang-tidy:
  - Disable misc-no-recursion: utf8_feed / resolve_path_depth /
    feed_utf8_byte are intentionally recursive state machines.
  - Disable clang-analyzer-valist.Uninitialized: false positive on
    console_printf (va_start IS called; tidy loses track across
    inline calls).
  - Disable clang-analyzer-core.UndefinedBinaryOperatorResult:
    false positive in ext2 dir-entry loop (static analyser path
    diverges from runtime).
  - Disable clang-analyzer-optin.performance.Padding: kernel structs
    (proc, ext2_volume_t) have fixed ABI layouts.
  - Disable bugprone-reserved-identifier / cert-dcl37-c /
    cert-dcl51-cpp: __c_ispeed / __c_ospeed are musl-compatible
    termios field names and cannot be renamed.

mk/repo.mk:
  - Add -Iuser/shell/include to lint compilation flags so that
    vega/*.h headers resolve and clang-tidy does not emit
    clang-diagnostic-error for every shell source file.

src/drivers/ata/ata.c:
  - CACHE_BLOCK_SECTORS → 8u, CACHE_BLOCK_BYTES → (u64)*512u to
    eliminate bugprone-implicit-widening-of-multiplication-result.

src/fs/vfs.c:
  - Add default: break to switch(kind) to fix
    bugprone-switch-missing-default-case.

user/apps/ncurses-hello/main.c:
  - Cast fputs() return to void (cert-err33-c; failure unrecoverable).

user/bin/cc.c:
  - Explicit (char **) cast for malloc result (bugprone-multi-level-
    implicit-pointer-conversion).
  - Cast fprintf() returns to void (cert-err33-c).

user/crt/alcor2_stdio_tty.c:
  - Cast setvbuf() returns to void (cert-err33-c; buffering is a
    best-effort hint).

user/shell/platform/fb_tty.c:
  - Explicit (size_t) casts in memmove() pointer arithmetic to fix
    bugprone-implicit-widening-of-multiplication-result.

user/shell/platform/io.c:
  - Explicit (char)(int) cast in digit accumulation to fix
    bugprone-narrowing-conversions.

+++ src/drivers/ata/ata.c
@@ -476,8 +476,8 @@ static i64 pio_write(ata_drive_t *d, u64 lba, u32 count, const void *buf)
  * fetch a full 4 KB block so adjacent reads land hot.
  */
 
-#define CACHE_BLOCK_SECTORS 8
-#define CACHE_BLOCK_BYTES   (CACHE_BLOCK_SECTORS * 512)
+#define CACHE_BLOCK_SECTORS 8u
+#define CACHE_BLOCK_BYTES   ((u64)CACHE_BLOCK_SECTORS * 512u)
 #define CACHE_NUM_ENTRIES   1024
 #define CACHE_INVALID_LBA   ((u64) - 1)
